+++ src/main/java/it/s3sync/listener/WatchListener.java
@@ -26,33 +26,25 @@ import org.slf4j.LoggerFactory;
 
 import com.sun.nio.file.SensitivityWatchEventModifier;
 
-import it.s3sync.listener.SynchronizationMessageDto.S3Action;
-import it.s3sync.listener.WatchListeners.Operation;
 import it.s3sync.service.SynchronizationService;
-import it.s3sync.service.UploadService;
-import it.s3sync.utils.FileUtils;
+import it.s3sync.utils.SpringContext;
 
 public class WatchListener implements Runnable {
 
-	private UploadService uploadService;
 	private String remoteFolder;
 	private String localRootFolder;
 	private WatchService watchService;
-	private SynchronizationService synchronizationService;
 
 	private Map<String, WatchKey> watchKeys = new HashMap<>();
 	private Set<String> directories = new HashSet<>();
 
 	private static final Logger logger = LoggerFactory.getLogger(WatchListener.class);
 
-	public WatchListener(UploadService uploadService, SynchronizationService synchronizationService,
-			String remoteFolder, String localRootFolder) {
+	public WatchListener(String remoteFolder, String localRootFolder) {
 		logger.info("[[INFO]] Constructing watchlistener for root local folder: {} with remote folder {}",
 				localRootFolder, remoteFolder);
-		this.uploadService = uploadService;
 		this.remoteFolder = remoteFolder;
 		this.localRootFolder = localRootFolder;
-		this.synchronizationService = synchronizationService;
 	}
 
 	@Override
@@ -176,106 +168,6 @@ public class WatchListener implements Runnable {
 	}
 
 	private void managingEvent(WatchEvent<?> event, Watchable watchable) {
-		String listenerPath = watchable.toString();
-		String resourceName = event.context().toString();
-		String fullLocation = listenerPath + "/" + resourceName;
-		Path fullPath = Path.of(fullLocation);
-
-		boolean workAsDirectory = Files.isDirectory(fullPath);
-		
-		Operation operation = new Operation();
-		operation.setS3Action("ENTRY_CREATE".equals(event.kind().name()) ? 
-				S3Action.CREATE : "ENTRY_MODIFY".equals(event.kind().name()) ? 
-						S3Action.MODIFY : S3Action.DELETE);
-		if(workAsDirectory) {
-			operation.setOnFile(fullLocation);
-		} else {
-			operation.setOnFile(fullLocation.replaceFirst(localRootFolder, ""));			
-		}
-		if (FileUtils.notMatchFilters(synchronizationService.getExclusionPattern(localRootFolder), fullLocation)) {
-			switch (event.kind().name()) {
-			case "ENTRY_CREATE":
-			case "ENTRY_MODIFY":
-				logger.debug("[[DEBUG]] Managing event CREATED on file {}", fullLocation);
-				if (workAsDirectory) {
-					logger.debug("[[DEBUG]] CREATE event on folder {}", fullLocation);
-					try {
-						Files.walkFileTree(fullPath, new SimpleFileVisitor<Path>() {
-							@Override
-							public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs)
-									throws IOException {
-								logger.debug("[[DEBUG]] Add watchKey on {}", fullLocation);
-								watchKeys.put(dir.toString(), dir.register(watchService, new WatchEvent.Kind[] {
-										StandardWatchEventKinds.ENTRY_CREATE, StandardWatchEventKinds.ENTRY_DELETE,
-										StandardWatchEventKinds.ENTRY_MODIFY, StandardWatchEventKinds.OVERFLOW },
-										SensitivityWatchEventModifier.MEDIUM));
-								logger.debug("[[DEBUG]] Add {} to folder tree", fullLocation);
-								directories.add(dir.toString());
-								return FileVisitResult.CONTINUE;
-							}
-						});
-					} catch (IOException e) {
-						logger.error("Exception", e);
-					}
-				} else {
-					logger.debug("[[DEBUG]] Managing event CREATE/MODIFY on file {}, start upload", fullLocation);
-					try {
-						uploadService.upload(fullPath, remoteFolder, fullLocation.replaceFirst(localRootFolder, ""),
-								Files.getLastModifiedTime(fullPath).toMillis());
-					} catch (Exception e) {
-						logger.error("Exception saving file {}", fullLocation, e);
-					}
-				}
-				break;
-			case "ENTRY_DELETE":
-				logger.debug("[[DEBUG]] Managing event DELETE on file {}", fullLocation);
-				if(workAsDirectory) {
-					if(watchKeys.containsKey(fullLocation)) {
-						if (directories.contains(fullLocation)) {
-							logger.debug("[[DEBUG]] DELETE event on folder {}", fullLocation);
-							if (watchKeys.containsKey(fullLocation)) {
-								logger.debug("[[DEBUG]] Kill watchKey and remote folder from tree");
-								watchKeys.get(fullLocation).cancel();
-								watchKeys.remove(fullLocation);
-							} else {
-								logger.warn("[[WARN]] WatchKey not found for this folder " + fullLocation);
-							}
-							logger.debug("[[DEBUG]] Delete all files in folder {}", fullLocation);
-							try {
-								String relativeLocation = fullLocation.replaceFirst(localRootFolder, "");
-								if (relativeLocation.isBlank()) {
-									synchronizationService.removeSynchronizationFolder(
-											synchronizationService.getSynchronizedLocalRootFolderByRemoteFolder(remoteFolder));
-								} else {
-									synchronizationService.addSynchronizationExclusionPattern(
-											synchronizationService.getSynchronizedLocalRootFolderByRemoteFolder(remoteFolder),
-											"^" + relativeLocation);
-								}
-								logger.debug("[[DEBUG]] Remove folder {} from the folders tree");
-								if(directories.contains(fullLocation)) {
-									directories.remove(fullLocation);									
-								}
-							} catch (Exception e) {
-								logger.error("Exception removing file {}", fullLocation, e);
-							}
-						}
-					} else {
-						logger.info("[[INFO]] Deleted folder was not in listening. Doing nothing...");
-					}
-				} else {
-					logger.debug("[[DEBUG]] DELETE event on file {}", fullLocation);
-					if (!uploadService.delete(remoteFolder, fullLocation.replaceFirst(localRootFolder, ""))) {
-						logger.error("[[ERROR]] Error deleting file {} from S3", fullLocation);
-					}
-				}
-				break;
-			default:
-				logger.debug("[[DEBUG]] Unhandled event {} on file {}", event.kind(), fullLocation);
-			}
-			logger.debug("[[DEBUG]] Finish serving event on {}", fullLocation);
-		} else {
-			logger.info("[[DEBUG]] Event on file {} was skipped by regexp filters.", fullLocation);
-		}
 	}
 
 	public void addFolder(String fullLocation) {
@@ -297,6 +189,42 @@ public class WatchListener implements Runnable {
 		}
 	}
 
+	public void justRemoteSyncFolder(String fullLocation) {
+		if(watchKeys.containsKey(fullLocation)) {
+			if (directories.contains(fullLocation)) {
+				logger.debug("[[DEBUG]] DELETE event on folder {}", fullLocation);
+				if (watchKeys.containsKey(fullLocation)) {
+					logger.debug("[[DEBUG]] Kill watchKey and remote folder from tree");
+					watchKeys.get(fullLocation).cancel();
+					watchKeys.remove(fullLocation);
+				} else {
+					logger.warn("[[WARN]] WatchKey not found for this folder " + fullLocation);
+				}
+				logger.debug("[[DEBUG]] Delete all files in folder {}", fullLocation);
+				try {
+					String relativeLocation = fullLocation.replaceFirst(localRootFolder, "");
+					if (relativeLocation.isBlank()) {
+						SpringContext.getBean(SynchronizationService.class).removeSynchronizationFolder(
+								SpringContext.getBean(SynchronizationService.class).getSynchronizedLocalRootFolderByRemoteFolder(remoteFolder));
+					} else {
+						SpringContext.getBean(SynchronizationService.class).addSynchronizationExclusionPattern(
+								SpringContext.getBean(SynchronizationService.class).getSynchronizedLocalRootFolderByRemoteFolder(remoteFolder),
+								"^" + relativeLocation);
+					}
+					logger.debug("[[DEBUG]] Remove folder {} from the folders tree");
+					if(directories.contains(fullLocation)) {
+						directories.remove(fullLocation);									
+					}
+				} catch (Exception e) {
+					logger.error("Exception removing file {}", fullLocation, e);
+				}
+			}
+		} else {
+			logger.info("[[INFO]] Deleted folder was not in listening. Doing nothing...");
+		}
+
+	}
+	
 	public void removeFolder(String fullLocation) {
 		synchronized (watchKeys) {
 			watchKeys.remove(fullLocation);
