+++ src/main/java/it/s3sync/listener/WatchListeners.java
@@ -9,9 +9,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import it.s3sync.listener.SynchronizationMessageDto.S3Action;
-import it.s3sync.service.SynchronizationService;
-import it.s3sync.service.UploadService;
+import it.s3sync.sync.EventData.S3Action;
 
 public class WatchListeners {
 	
@@ -23,10 +21,8 @@ public class WatchListeners {
 	
 	private static final Logger logger = LoggerFactory.getLogger(WatchListeners.class);
 
-	public static void startThread(UploadService uploadService, SynchronizationService synchronizationService,
-			String remoteFolder, String localRootFolder) {
-		WatchListener listener = new WatchListener(uploadService, synchronizationService, remoteFolder,
-				localRootFolder);
+	public static void startThread(String remoteFolder, String localRootFolder) {
+		WatchListener listener = new WatchListener(remoteFolder, localRootFolder);
 		Thread thread = new Thread(listener);
 		thread.start();
 		logger.info("[[INFO]] Starting watch thread {} on local/remote folders {} -> {}", thread.getName(),
@@ -71,6 +67,14 @@ public class WatchListeners {
 		return WatchListeners.threadSemaphore.get() == 0;
 	}
 	
+	public static void registerNewFolder(String localRootFolder, String folderFullPath) {
+		((WatchListener)threadPool.get(localRootFolder).getR()).addFolder(folderFullPath);
+	}
+	
+	public static void removeFolder(String localRootFolder, String folderFullPath) {
+		((WatchListener)threadPool.get(localRootFolder).getR()).justRemoteSyncFolder(folderFullPath);
+	}
+
 	public static class ThreadAndRunnable {
 		
 		public ThreadAndRunnable(Thread t, Runnable r) {
